卡牌61优化

+++ Core/GObject/GObjectDBExecHelper.h
@@ -1232,7 +1232,7 @@ struct DBFairySpar
 struct DBCollectCnt
 {
     UInt64 playerId;
-    UInt8 level;
+    UInt16 level;
     UInt16 bluecnt;
     UInt16 purlecnt;
     UInt16 orangecnt;
@@ -1625,7 +1625,7 @@ struct DBCard
 struct DBCardSuit
 {
     UInt64 playerId;
-    UInt8 id;
+    UInt16 id;
     UInt8 suit_mark;
     UInt8 active;
     UInt32 spe_mark;
@@ -3190,7 +3190,7 @@ SPECIALBEGIN(GObject::DBCollectCnt)
 SPECIALDEF(5)
 (
     UInt64, playerId,
-    UInt8, level,
+    UInt16, level,
     UInt16, bluecnt,
     UInt16, purlecnt,
     UInt16, orangecnt
@@ -3705,7 +3705,7 @@ SPECIALBEGIN (GObject::DBCardSuit)
 SPECIALDEF(6)
     (
     UInt64, playerId,
-    UInt8, id,
+    UInt16, id,
     UInt8, suit_mark,
     UInt8, active,
     UInt32, spe_mark,
